[CS-1890] - Add PayMerch hook to b able to test
